requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
gunicorn==20.1.0
h11==0.13.0
html5lib==1.1
idna==3.3
lxml==4.9.1
mysql-connector==2.2.9
mysql-connector-python==8.0.30
numpy==1.23.1
outcome==1.2.0
pandas==1.4.3
protobuf==3.20.1
psycopg2==2.9.3
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.11
webencodings==0.5.1
wsproto==1.1.0
